Create generic detail fragment class
- Move generic code of alert detail fragment class to detail fragment class
